requirements.txt
antlr4-python3-runtime==4.9.3
calc_assoc_cuda==0.0.0
calc_pixel_feats_cuda==0.0.0
calc_spixel_feats_cuda==0.0.0
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
fvcore==0.1.5.post20221221
hydra-core==1.3.2
imageio==2.37.0
init_spixel_feats_cuda==0.0.0
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.0
nvidia-cublas-cu12 @ file:///home/pnxc/nvidia_cublas_cu12-12.8.3.14-py3-none-manylinux_2_27_x86_64.whl#sha256=3f0e05e7293598cf61933258b73e66a160c27d59c4422670bf0b79348c04be44
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12 @ file:///home/pnxc/nvidia_cudnn_cu12-9.7.1.26-py3-none-manylinux_2_27_x86_64.whl#sha256=6d011159a158f3cfc47bf851aea79e31bcff60d530b70ef70474c84cac484d07
nvidia-cufft-cu12 @ file:///home/pnxc/nvidia_cufft_cu12-11.3.3.41-py3-none-manylinux2014_x86_64.manylinux_2_17_x86_64.whl#sha256=da650080ab79fcdf7a4b06aa1b460e99860646b176a43f6208099bdc17836b6a
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12 @ file:///home/pnxc/nvidia_curand_cu12-10.3.9.55-py3-none-manylinux_2_27_x86_64.whl#sha256=8387d974240c91f6a60b761b83d4b2f9b938b7e0b9617bae0f0dafe4f5c36b86
nvidia-cusolver-cu12 @ file:///home/pnxc/nvidia_cusolver_cu12-11.7.2.55-py3-none-manylinux_2_27_x86_64.whl#sha256=4d1354102f1e922cee9db51920dba9e2559877cf6ff5ad03a00d853adafb191b
nvidia-cusparse-cu12 @ file:///home/pnxc/nvidia_cusparse_cu12-12.5.7.53-py3-none-manylinux2014_x86_64.manylinux_2_17_x86_64.whl#sha256=3c1b61eb8c85257ea07e9354606b26397612627fdcd327bfd91ccf6155e7c86d
nvidia-cusparselt-cu12 @ file:///home/pnxc/nvidia_cusparselt_cu12-0.6.3-py3-none-manylinux2014_x86_64.whl#sha256=e5c8a26c36445dd2e6812f1177978a24e2d37cacce7e090f297a688d1ec44f46
nvidia-nccl-cu12 @ file:///home/pnxc/nvidia_nccl_cu12-2.26.2-py3-none-manylinux2014_x86_64.manylinux_2_17_x86_64.whl#sha256=694cf3879a206553cc9d7dbda76b13efaf610fdb70a50cba303de1b0d1530ac6
nvidia-nvjitlink-cu12 @ file:///home/pnxc/nvidia_nvjitlink_cu12-12.8.61-py3-none-manylinux2010_x86_64.manylinux_2_12_x86_64.whl#sha256=45fd79f2ae20bd67e8bc411055939049873bfd8fac70ff13bd4865e0b9bdab17
nvidia-nvtx-cu12==12.8.55
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
pycocotools==2.0.10
pyparsing==3.2.3
pysodmetrics==1.4.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rel_to_abs_index_cuda==0.0.0
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
smear_cuda==0.0.0
spectral==0.24
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
torch @ file:///home/pnxc/torch-2.7.1%2Bcu128-cp310-cp310-manylinux_2_28_x86_64.whl#sha256=d6c3cba198dc93f93422a8545f48a6697890366e4b9701f54351fc27e2304bd3
torchprofile==0.0.4
torchsummary==1.5.1
torchvision==0.22.1
tqdm==4.67.1
triton @ file:///home/pnxc/triton-3.3.1-cp310-cp310-manylinux_2_27_x86_64.manylinux_2_28_x86_64.whl#sha256=b74db445b1c562844d3cfad6e9679c72e93fdfb1a90a24052b03bb5c49d1242e
typing_extensions==4.14.1
yacs==0.1.8
